[Maybe-working-Checkpoint] Refactor LoadingAnimation logic
