--- src/views/loading/loading.scss
@@ -1,6 +1,0 @@
-.peak-loader {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-}
