+++ src/views/loading/react-lottie-container/react-lottie-container.scss
@@ -1,0 +1,6 @@
+.peak-react-lottie-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+}
